Global Chat (#126)
* [front]
- rename chat to private-chat
- clone private-chat to global-chat
- move interfaces and service into pages
- ChatRoomI add default type
- add RoomType
- list available room and joined room
- split in 2 column with 1 row, for creating and joining
- reset on join
  - clear form
  - unselect room
  - remove available room
- change budget

* [WS]
- working global chat, each PROTECTED and PUBLIC
  chat can be join according to type
- split popup in two types, create-join and management
- add new event NewUserJoin
- base for management
  - add two space, one for edit details and another to manage users
  - disable changement for none owner
- add new event, NewDetailsRoom
- remove sensible information in data transmited
- add roomAction to kick ban mute etc ...
  - add roomAction event KICK

* [back]
- format
- remove already joined from available room
- add check for already joined room
- fix remove already joined
- fix available room
- fix joined room data missing user
- add changeRoom details

* [other]
- errcomponent permission

+++ src/nestjs/src/modules/database/chatRoom/service.ts
@@ -26,34 +26,26 @@ export class DBChatRoomService {
 
 	async returnOne(chatId: number) {
 		const tmp = await this.chatRoomRepo.findOneBy({ id: chatId });
-		if (tmp)
-			return tmp;
-		else
-			throw new NotFoundException("ChatRoom not found");
+		if (tmp) return tmp;
+		else throw new NotFoundException("ChatRoom not found");
 	}
 
 	async update(chatId: number, post: DBChatRoomPost) {
 		const tmp = await this.chatRoomRepo.findOneBy({ id: chatId });
-		if (tmp)
-			return await this.chatRoomRepo.update(chatId, post);
-		else
-			throw new NotFoundException("ChatRoom not found");
+		if (tmp) return await this.chatRoomRepo.update(chatId, post);
+		else throw new NotFoundException("ChatRoom not found");
 	}
 
 	async updateType(chatId: number, post: DBChatRoomTypePost) {
 		const tmp = await this.chatRoomRepo.findOneBy({ id: chatId });
-		if (tmp)
-			return await this.chatRoomRepo.update(chatId, post);
-		else
-			throw new NotFoundException("ChatRoom not found");
+		if (tmp) return await this.chatRoomRepo.update(chatId, post);
+		else throw new NotFoundException("ChatRoom not found");
 	}
 
 	async delete(chatId: number) {
 		const tmp = await this.chatRoomRepo.findOneBy({ id: chatId });
-		if (tmp)
-			return await this.chatRoomRepo.delete(chatId);
-		else
-			throw new NotFoundException("ChatRoom not found");
+		if (tmp) return await this.chatRoomRepo.delete(chatId);
+		else throw new NotFoundException("ChatRoom not found");
 	}
 
 	async getAllPrivateRoom(chatId: number): Promise<ChatRoomEntity> {
