+++ src/nestjs/src/modules/database/friendRequest/service.ts
@@ -19,10 +19,9 @@ export class DBFriendRequestService {
 	) {}
 
 	async create(post: DBFriendRequestPost, meId: number) {
-		const user1 = await this.userRepo.findOneBy({id: meId});
-		const user2 = await this.userRepo.findOneBy({id: post.friendId});
-		if (!user1 || !user2)
-			throw new NotFoundException("User not found");
+		const user1 = await this.userRepo.findOneBy({ id: meId });
+		const user2 = await this.userRepo.findOneBy({ id: post.friendId });
+		if (!user1 || !user2) throw new NotFoundException("User not found");
 		let request = new FriendRequestEntity();
 		request.meId = meId;
 		request.friendId = post.friendId;
@@ -35,37 +34,40 @@ export class DBFriendRequestService {
 	}
 
 	async returnOne(me_id: number, friendId: number) {
-		const tmp = await this.friendRequestRepo.findOneBy({ meId: me_id, friendId: friendId });
-		if (tmp)
-			return tmp;
-		else
-			throw new NotFoundException("FriendRequest relation not found");
+		const tmp = await this.friendRequestRepo.findOneBy({
+			meId: me_id,
+			friendId: friendId,
+		});
+		if (tmp) return tmp;
+		else throw new NotFoundException("FriendRequest relation not found");
 	}
 
 	async delete(me_id: number, friendId: number) {
-		const tmp = await this.friendRequestRepo.findOneBy({ meId: me_id, friendId: friendId });
-		if (tmp) 
-			return await this.friendRequestRepo.delete(tmp);
-		else 
-			throw new NotFoundException("FriendRequest relation not found");
+		const tmp = await this.friendRequestRepo.findOneBy({
+			meId: me_id,
+			friendId: friendId,
+		});
+		if (tmp) return await this.friendRequestRepo.delete(tmp);
+		else throw new NotFoundException("FriendRequest relation not found");
 	}
 
 	async acceptReq(me_id: number, friendId: number) {
-		const req = await this.friendRequestRepo.findOneBy({ meId: me_id, friendId: friendId });
-		if (req)
-		{
-			await this.friendService.create({friendId: friendId}, me_id);
+		const req = await this.friendRequestRepo.findOneBy({
+			meId: me_id,
+			friendId: friendId,
+		});
+		if (req) {
+			await this.friendService.create({ friendId: friendId }, me_id);
 			await this.friendRequestRepo.delete(req);
-		}
-		else
-			throw new NotFoundException("FriendRequest relation not found");
+		} else throw new NotFoundException("FriendRequest relation not found");
 	}
 
 	async rejectReq(me_id: number, friendId: number) {
-		const req = await this.friendRequestRepo.findOneBy({ meId: me_id, friendId: friendId });
-		if (req)
-			await this.friendRequestRepo.delete(req);
-		else
-			throw new NotFoundException("FriendRequest relation not found");
+		const req = await this.friendRequestRepo.findOneBy({
+			meId: me_id,
+			friendId: friendId,
+		});
+		if (req) await this.friendRequestRepo.delete(req);
+		else throw new NotFoundException("FriendRequest relation not found");
 	}
 }
