+++ src/nestjs/src/modules/database/userChatRoom/service.ts
@@ -4,7 +4,7 @@ import { InjectRepository } from "@nestjs/typeorm";
 
 import { UserChatRoomEntity } from "./entity";
 import { UserEntity } from "../user/entity";
-import { ChatRoomEntity } from "../chatRoom/entity";
+import { ChatRoomEntity, RoomType } from "../chatRoom/entity";
 
 import { DBUserChatRoomPost } from "./dto";
 
@@ -22,15 +22,11 @@ export class DBUserChatRoomService {
 	async create(post: DBUserChatRoomPost, userId: number, roomId: number) {
 		const userChat = new UserChatRoomEntity();
 		let user = await this.userRepo.findOneBy({ id: userId });
-		if (user) 
-			userChat.userId = userId;
-		else 
-			throw new NotFoundException("User not found");
+		if (user) userChat.userId = userId;
+		else throw new NotFoundException("User not found");
 		let room = await this.chatRoomRepo.findOneBy({ id: roomId });
-		if (room)
-			userChat.roomId = roomId;
-		else 
-			throw new NotFoundException("ChatRoom not found");
+		if (room) userChat.roomId = roomId;
+		else throw new NotFoundException("ChatRoom not found");
 		userChat.isOwner = post.isOwner;
 		userChat.isAdmin = post.isAdmin;
 		return await this.userChatRoomRepo.save(userChat);
@@ -45,10 +41,22 @@ export class DBUserChatRoomService {
 			userId: user,
 			roomId: room,
 		});
-		if (tmp) 
-			return tmp;
-		else 
-			throw new NotFoundException("userChatRoom not found");
+		if (tmp) return tmp;
+		else throw new NotFoundException("userChatRoom not found");
+	}
+
+	async returnOneWithUser(user: number, room: number) {
+		const tmp = await this.userChatRoomRepo.findOne({
+			relations: {
+				user: true
+			},
+			where: {
+				userId: user,
+				roomId: room,
+			},
+		});
+		if (tmp) return tmp;
+		else throw new NotFoundException("userChatRoom not found");
 	}
 
 	async update(user: number, room: number, post: DBUserChatRoomPost) {
@@ -56,10 +64,8 @@ export class DBUserChatRoomService {
 			userId: user,
 			roomId: room,
 		});
-		if (tmp) 
-			return await this.userChatRoomRepo.update(tmp, post);
-		else
-			throw new NotFoundException("userChatRoom not found");
+		if (tmp) return await this.userChatRoomRepo.update(tmp, post);
+		else throw new NotFoundException("userChatRoom not found");
 	}
 
 	async delete(user: number, room: number) {
@@ -67,10 +73,8 @@ export class DBUserChatRoomService {
 			userId: user,
 			roomId: room,
 		});
-		if (tmp)
-			return await this.userChatRoomRepo.delete(tmp);
-		else
-			throw new NotFoundException("userChatRoom not found");
+		if (tmp) return await this.userChatRoomRepo.delete(tmp);
+		else throw new NotFoundException("userChatRoom not found");
 	}
 
 	async getAllPrivateRoom(userId: number): Promise<ChatRoomEntity[]> {
@@ -87,6 +91,7 @@ export class DBUserChatRoomService {
 						id: userId,
 					},
 				},
+				type: RoomType.PRIVATE,
 			},
 			order: {
 				message: {
@@ -118,6 +123,132 @@ export class DBUserChatRoomService {
 					},
 				},
 			},
+		});
+	}
+
+	async getAllGlobalUserRoom(room_id: number): Promise<UserChatRoomEntity[]> {
+		return await this.userChatRoomRepo.find({
+			relations: {
+				user: true,
+				room: {
+					message: {
+						user: true,
+					},
+				},
+			},
+			where: [
+				{
+					roomId: room_id,
+					room: {
+						type: RoomType.PROTECTED,
+					},
+				},
+				{
+					roomId: room_id,
+					room: {
+						type: RoomType.PUBLIC,
+					},
+				},
+			],
+			order: {
+				room: {
+					message: {
+						updateAt: "ASC",
+					},
+				},
+			},
+		});
+	}
+
+	async getAllAvailableGlobalRoom(): Promise<ChatRoomEntity[]> {
+		return await this.chatRoomRepo.find({
+			relations: {
+				roomInfo: {
+					user: true,
+				},
+			},
+			where: [
+				{
+					type: RoomType.PUBLIC,
+				},
+				{
+					type: RoomType.PROTECTED,
+				},
+			],
+		});
+	}
+
+	async getAvailableGlobalRoom(room_id: number): Promise<ChatRoomEntity> {
+		return await this.chatRoomRepo.findOne({
+			relations: {
+				roomInfo: {
+					user: true,
+				},
+			},
+			where: [
+				{
+					id: room_id,
+					type: RoomType.PUBLIC,
+				},
+				{
+					id: room_id,
+					type: RoomType.PROTECTED,
+				},
+			],
+		});
+	}
+
+	async getJoinedGlobalRoom(room_id: number): Promise<ChatRoomEntity> {
+		return await this.chatRoomRepo.findOne({
+			relations: {
+				roomInfo: {
+					user: true,
+				},
+				message: {
+					user: true,
+				},
+			},
+			where: [
+				{
+					id: room_id,
+					type: RoomType.PUBLIC,
+				},
+				{
+					id: room_id,
+					type: RoomType.PROTECTED,
+				},
+			],
+		});
+	}
+
+	async getAllJoinedGlobalRoom(user_id: number): Promise<ChatRoomEntity[]> {
+		return await this.chatRoomRepo.find({
+			relations: {
+				roomInfo: {
+					user: true,
+				},
+				message: {
+					user: true,
+				},
+			},
+			where: [
+				{
+					roomInfo: {
+						user: {
+							id: user_id,
+						},
+					},
+					type: RoomType.PUBLIC,
+				},
+				{
+					roomInfo: {
+						user: {
+							id: user_id,
+						},
+					},
+					type: RoomType.PROTECTED,
+				},
+			],
 		});
 	}
